move substrate users db to provision pkg
so it can be reused
